Adding new worldgroup feature
This commit adds the worldgroups to the config option. By specifying
groups: [list of worlds] you can limit the flow of animals to-and-from
worlds. By default, if the teleporting-to world is the same as the world
teleported-from, it will be true; however if the worlds are different
you must have them in the group to allow the teleport!