Adding config flag for unknown teleport causes
In versions below 1.19.3, leaving a vehicle triggered an unknown teleport cause. If you are on versions below 1.19.3, setting this config value to FALSE should help catch this issue and remove the unnecessary player notification about an entity being left behind. In versions greater than 1.19.3, a DISMOUNT cause was added due to this report. This config value can safely be ignored on versions >= 1.19.3. Fixes #7