Modifying entity checking logic
Updated the logic of the entity gatherer when a player teleports.

We now will check if the entity is leashed or not first, and then proceed to remaining checks.

If the entity was not leashed, we will proceed to check if it was tamed. If it was tamed, and not sitting, we will proceed.

If neither of the above conditions were met, we will notify player.